+++ polymorphia-frontend/components/hall-of-fame/general/image/index.tsx
