+++ polymorphia-frontend/components/hall-of-fame/general/pagination/index.css
@@ -1,0 +1,6 @@
+@import "tailwindcss";
+@import "../../../../styles/globals.css";
+
+.pagination-wrapper {
+  @apply hall-of-fame-pagination-wrapper;
+}
