+++ polymorphia-frontend/components/hall-of-fame/general/top-bar/index.css
@@ -1,0 +1,10 @@
+@import "tailwindcss";
+@import "../../../../styles/globals.css";
+
+.top-bar {
+  @apply flex w-full gap-x-2 justify-between overflow-x-hidden;
+}
+
+.top-bar-buttons {
+  @apply flex gap-x-2 lg:gap-x-4;
+}
