+++ polymorphia-frontend/components/hall-of-fame/index.css
@@ -1,0 +1,6 @@
+@import "tailwindcss";
+@import "../../styles/globals.css";
+
+.hall-of-fame-loading-wrapper {
+  @apply relative min-h-[calc(100dvh-10rem)] flex flex-col flex-1 w-full h-full lg:min-h-20;
+}
