+++ polymorphia-frontend/components/hall-of-fame/mobile/main/index.css
@@ -1,0 +1,12 @@
+@import "tailwindcss";
+@import "../../../../styles/globals.css";
+
+.main-mobile {
+  @apply w-full px-2 opacity-0 flex flex-col h-full;
+}
+
+.main-mobile-search-wrapper {
+  @apply hall-of-fame-search-wrapper min-h-10;
+  scrollbar-gutter: stable both-edges;
+  overflow-y: hidden;
+}
