+++ polymorphia-frontend/styles/globals.css
@@ -67,6 +67,22 @@
   @apply duration-400 ease-[cubic-bezier(0.34,1,0.2,1)];
 }
 
+@utility hall-of-fame-rank-wrapper {
+  @apply w-full overflow-y-scroll custom-scrollbar;
+}
+
+@utility hall-of-fame-pagination-wrapper {
+  @apply w-full h-14 flex items-center text-2xl py-2 lg:px-3;
+}
+
+@utility hall-of-fame-record-wrapper {
+  @apply w-full rounded-xl shadow-lg bg-primary-light dark:bg-primary-dark;
+}
+
+@utility hall-of-fame-search-wrapper {
+  @apply w-full flex justify-between mb-3;
+}
+
 @source inline('grid-rows-1 grid-rows-2 grid-rows-3 grid-rows-4 grid-cols-1 grid-cols-2 grid-cols-3 grid-cols-4');
 
 input:-webkit-autofill,
