--- polymorphia-frontend/views/hall-of-fame/general/error/index.css
@@ -1,12 +1,0 @@
-@import "tailwindcss";
-@import "../../../../styles/globals.css";
-
-@tailwind utilities;
-
-.error-no-data-wrapper {
-  @apply m-auto;
-
-  h1 {
-    @apply text-3xl text-center lg:text-4xl;
-  }
-}
