--- polymorphia-frontend/views/hall-of-fame/general/image/index.css
@@ -1,16 +1,0 @@
-@import "tailwindcss";
-@import "../../../../styles/globals.css";
-
-@tailwind utilities;
-
-.image-wrapper {
-  @apply h-full relative aspect-square;
-}
-
-.image-img {
-  @apply object-cover rounded-xl lg:rounded-bl-xl lg:rounded-tr-none lg:rounded-br-none;
-}
-
-.image-badge {
-  @apply rounded-tl-xl rounded-br-xl w-5 text-base lg:text-2xl lg:min-w-8 lg:rounded-br-none;
-}
