--- polymorphia-frontend/views/hall-of-fame/general/index.css
@@ -1,24 +1,0 @@
-@import "tailwindcss";
-@import "../../../styles/globals.css";
-
-@tailwind utilities;
-
-@utility hall-of-fame-rank-wrapper {
-  @apply w-full overflow-y-scroll custom-scrollbar;
-}
-
-@utility hall-of-fame-pagination-wrapper {
-  @apply w-full h-14 flex items-center text-2xl py-2 lg:px-3;
-}
-
-@utility hall-of-fame-record-wrapper {
-  @apply w-full rounded-xl shadow-lg bg-primary-light dark:bg-primary-dark;
-}
-
-@utility hall-of-fame-search-wrapper {
-  @apply w-full flex justify-between mb-3;
-}
-
-.hall-of-fame-loading-wrapper {
-  @apply relative min-h-[calc(100dvh-10rem)] flex flex-col flex-1 w-full h-full lg:min-h-20;
-}
