--- polymorphia-frontend/views/hall-of-fame/general/list/index.css
@@ -1,16 +1,0 @@
-@import "tailwindcss";
-@import "../../../../styles/globals.css";
-@import "../index.css";
-
-@tailwind utilities;
-
-.list-desktop-rank-wrapper {
-  @apply hall-of-fame-rank-wrapper flex flex-col flex-1 min-h-0 overflow-y-auto gap-y-2 2xl:gap-y-4;
-}
-
-.list-mobile-rank-wrapper {
-  @apply w-full my-auto grid grid-cols-1 gap-3 min-[650px]:grid-cols-2 overflow-y-auto custom-scrollbar;
-  scrollbar-gutter: stable both-edges;
-}
-
-@source inline('grid-cols-1 grid-cols-2');
