--- polymorphia-frontend/views/hall-of-fame/general/pagination/index.css
@@ -1,11 +1,0 @@
-@import "tailwindcss";
-@import "../../../../styles/globals.css";
-@import "../index.css";
-
-@tailwind utilities;
-
-.pagination-wrapper {
-  @apply hall-of-fame-pagination-wrapper;
-}
-
-@source inline('justify-end min-h-14');
