--- polymorphia-frontend/views/hall-of-fame/general/points-summary/index.css
@@ -1,18 +1,0 @@
-@import "tailwindcss";
-@import "../../../../styles/globals.css";
-
-@tailwind utilities;
-
-.points-summary {
-  @apply w-24 h-full flex flex-col-reverse items-end justify-center;
-  @apply lg:items-center lg:flex-col;
-  @apply lg:bg-[#d4d4d4] lg:dark:bg-secondary-dark lg:rounded-tr-xl lg:rounded-br-xl;
-
-  h1 {
-    @apply text-3xl lg:text-4xl;
-  }
-
-  h3 {
-    @apply text-xl lg:text-2xl;
-  }
-}
