--- polymorphia-frontend/views/hall-of-fame/mobile/main/index.css
@@ -1,17 +1,0 @@
-@import "tailwindcss";
-@import "../../../../styles/globals.css";
-@import "../../general/index.css";
-
-@tailwind utilities;
-
-.main-mobile {
-  @apply w-full px-2 opacity-0 flex flex-col h-full;
-}
-
-.main-mobile-search-wrapper {
-  @apply hall-of-fame-search-wrapper min-h-10;
-  scrollbar-gutter: stable both-edges;
-  overflow-y: hidden;
-}
-
-@source inline('grid-cols-1 grid-cols-2');
