--- polymorphia-frontend/views/hall-of-fame/mobile/main/index.tsx
@@ -1,22 +1,0 @@
-"use client";
-
-import "./index.css";
-import "../../general/index.css";
-import { useScaleShow } from "@/animations/ScaleShow";
-import HallOfFamePagination from "@/views/hall-of-fame/general/pagination";
-import HallOfFameList from "@/views/hall-of-fame/general/list";
-import HallOfFameTopBar from "@/views/hall-of-fame/general/top-bar";
-
-export default function HallOfFameMobile() {
-  const wrapperRef = useScaleShow();
-
-  return (
-    <div ref={wrapperRef} className="main-mobile">
-      <div className="main-mobile-search-wrapper">
-        <HallOfFameTopBar />
-      </div>
-      <HallOfFameList />
-      <HallOfFamePagination />
-    </div>
-  );
-}
